+++ src/compiler/checker/mod.rs
@@ -1,5 +1,7 @@
-mod lines_0_8000;
-pub use lines_0_8000::*;
+mod lines_0_4000;
+pub use lines_0_4000::*;
+mod lines_4000_8000;
+pub use lines_4000_8000::*;
 mod lines_8000_16000;
 pub use lines_8000_16000::*;
 mod lines_16000_24000;
